classification_misc

+++ src/datamodules/datasets.py
@@ -315,7 +315,7 @@ class BreastPairedDataset(BaseDataset):
         mlo_img = self.read_dcm(mlo_path)
         target = research_df[self.target_col].iloc[0]
         if type(target) == str:
-            target = torch.tensor(self.ClasName2Index[target])
+            target = torch.tensor(self.ClasName2Index[target], dtype=torch.long)
         target = torch.tensor(target)
         if self.transforms is not None:
             augmentation = self.transforms(image=cc_img)
